chore: refactor cluster page
