+++ frontend/src/pages/clusters/add-new-cluster.test.tsx
@@ -1,0 +1,9 @@
+import { it, describe, vi } from "vitest";
+import { render } from "@testing-library/react";
+import { AddNewCluster } from "./add-new-cluster";
+
+describe("addNewCluster", () => {
+  it("should render", () => {
+    render(<AddNewCluster onSubmit={vi.fn()} />);
+  });
+});
